cleanup: int -> size_t since there is no actual signed usage

+++ address.h
@@ -5,11 +5,11 @@
 
 #include "status.h"
 
-bool addr_rr(Status *const status, Regex *const start, Regex *const end, const int id);
-bool addr_rn(Status *const status, Regex *const start, const int end, const int id);
-bool addr_nr(Status *const status, const int start, Regex *const end, const int id);
-bool addr_nn(Status *const status, const int start, const int end, const int id);
+bool addr_rr(Status *const status, Regex *const start, Regex *const end, const size_t id);
+bool addr_rn(Status *const status, Regex *const start, const size_t end, const size_t id);
+bool addr_nr(Status *const status, const size_t start, Regex *const end, const size_t id);
+bool addr_nn(Status *const status, const size_t start, const size_t end, const size_t id);
 bool addr_r(Status *const status, Regex *const regex);
-bool addr_n(const Status *const status, const int line);
+bool addr_n(const Status *const status, const size_t line);
 
 #endif /* ADDRESS_H */
